Configuracion general terminada

+++ src/main/java/com/proyectospa/spa_app/repository/TurnoRepository.java
@@ -14,7 +14,6 @@ public interface TurnoRepository extends JpaRepository<Turno, Integer> {
     List<Turno> findByProfesionalIdAndEstado(Integer profesionalId, EstadoTurno estado);
     List<Turno> findByProfesionalId(Integer profesionalId);
     List<Turno> findByClienteId(Integer clienteId);
-    List<Turno> findByClienteIdOrderByFechaDesc(Long clienteId);
+    List<Turno> findByClienteIdOrderByFechaDesc(Integer clienteId);
     List<Turno> findByProfesionalIdAndClienteId(Integer profesionalId, Integer clienteId);
-    
 }
